+++ inc/dg/topology/functions.h
@@ -65,19 +65,6 @@ DG_DEVICE static inline float zero( float x, float y) {return 0.;}
 ///@brief \f[ f(x,y,z) = 0\f]
 DG_DEVICE static inline float zero( float x, float y, float z) {return 0.;}
 
-///@brief \f[ f(x) = x\f]
-DG_DEVICE static inline float cooX1d( float x) {return x;}
-///@brief \f[ f(x,y) = x\f]
-DG_DEVICE static inline float cooX2d( float x, float y) {return x;}
-///@brief \f[ f(x,y,z) = x\f]
-DG_DEVICE static inline float cooX3d( float x, float y, float z) {return x;}
-
-///@brief \f[ f(x,y) = y\f]
-DG_DEVICE static inline float cooY2d( float x, float y) {return y;}
-///@brief \f[ f(x,y,z) = y\f]
-DG_DEVICE static inline float cooY3d( float x, float y, float z) {return y;}
-///@brief \f[ f(x,y,z) = z\f]
-DG_DEVICE static inline float cooZ3d( float x, float y, float z) {return z;}
 } //namespace dg
 
 ///@}